[ET-VK] Organize `StringUtil.h` and `Utils.h` into `utils/` and `VkUtils.h`
Everything in `utils/` will take the `namespace vkcompute:::utils`; everything in `VkUtils.h` deals with the Vulkan API directly.

Differential Revision: [D59281548](https://our.internmc.facebook.com/intern/diff/D59281548/)

[ghstack-poisoned]
