+++ backends/vulkan/runtime/api/Adapter.h
@@ -14,7 +14,7 @@
 
 #include <executorch/backends/vulkan/runtime/api/Pipeline.h>
 #include <executorch/backends/vulkan/runtime/api/Shader.h>
-#include <executorch/backends/vulkan/runtime/api/Utils.h>
+#include <executorch/backends/vulkan/runtime/api/utils/VecUtils.h>
 
 #include <executorch/backends/vulkan/runtime/api/memory/Allocator.h>
