+++ backends/vulkan/runtime/api/memory/Buffer.h
@@ -12,7 +12,7 @@
 
 #include <executorch/backends/vulkan/runtime/api/vk_api.h>
 
-#include <executorch/backends/vulkan/runtime/api/Utils.h>
+#include <executorch/backends/vulkan/runtime/api/utils/VecUtils.h>
 
 #include <executorch/backends/vulkan/runtime/api/memory/vma_api.h>
